Move actions into a new component
